requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
python-multipart==0.0.9
python-dotenv==1.0.1